ci: prevent PR merges when analysis is incomplete
This change takes advantage of the new `--fail-incomplete` flag to be
more strict in how the Phylum analysis check is handled. Before this
change it was possible to merge a PR with new/changed dependencies even
before the Phylum analysis was complete. With this change, the status
check will report as a failure, blocking the ability to merge, when the
analysis is incomplete. This provides a degree of strictness to prevent
adding *any* new dependency until it passes established policy.