feat!: remove `SUCCESS` comments from PRs/MRs (#526)
This change removes the `SUCCESS` comments from pull requests (PRs) and
merge requests (MRs). Other comment types are unchanged. This change is
necessary because the `SUCCESS` comments are considered by existing
customers to be too noisy and not useful since the related status check
will show as passing when the analysis is successful. Plus, the logs
contain the same information as was included in the `SUCCESS` comments.

Closes #78

BREAKING CHANGE: PRs/MRs will no longer have `SUCCESS` comments.