docs: update language for Veracode transition
This change updates the `@phylum.io` emails to `@veracode.com` ones.

It also updates the docs, help output, and error messages to no longer
make a distinction between the community edition and the pro edition of
Phylum now that there is no community edition and all (paid) users have
the same features.