ci: restrict `rich-codex` dependency range
This change restricts the `rich-codex` dependency from using a version
that does not work to validate configs. A regression was added to the
tool in the v1.2.9 release and has been fixed with a PR that has been
merged but not yet included in a release. This is a "restriction"
instead of a pin because the two versions that are restricted are known
to be non-functional while previous and future versions are functional.