build: update `poetry` to latest version (#554)
Poetry is the workflow management tool used for this project and forms
the root of all other actions taken when working with this repository.
It is also used to manage dependencies and therefore should be treated
very carefully, with updates to newer versions taken deliberately.

This PR updates `poetry` to the latest version of v2.1.2 to account for
the major [changes introduced](https://python-poetry.org/history) in
crossing from v1.x to v2.x, with these actions taken:

* Bump all instances of `poetry` to the new version
  * Installs in workflows
  * pre-commit hook revision
  * Dockerfiles
  * Full comparison:
    * https://github.com/python-poetry/poetry/compare/1.8.5...2.1.2
* Bump the `poetry-core` build system version in `pyproject.toml`
  * v2.1.2 matches Poetry version and adopts matching breaking changes
  * Full comparison:
    * https://github.com/python-poetry/poetry-core/compare/1.9.1...2.1.2
* Update the lockfile with the new version of `poetry`

These updates added changes and new features that required updates to
the use of `poetry` in this project. The changes made include:

* `poetry check` instances now include `--strict` option
* `poetry lock` instances no longer have `--no-update` option
* `poetry install` instances are now `poetry sync`
* Update `poetry-check` pre-commit hook options
* Update `pyproject.toml` to match new expected PEP 621 compliant format

Other changes made include:

* Update documentation
  * `CONTRIBUTING.md` was brought current
* Add company details to `LICENSE`
* Update tests
  * Adjust for new expected package metadata and it's related format
  * Test version ranges with `SpecifierSet` from the `packaging` package
  * Add type hints