build: revert breaking PSR update (#566)
This commit mostly seeks to revert the change made in #563, where the
Python Semantic Release (PSR) dependency was updated to a new major
version. This version introduces a lot of breaking changes and should
not have happened. There was a comment above the previous pin stating as
much, but there was no matching Dependabot configuration, which is how
the PR was introduced. That configuration was added, along with a TODO
statement about migrating to the latest major version.

Additionally, the `rich-codex` dependency restriction was updated to
account for the v1.2.11 release where a regression affecting this repo
was fixed. Finally, the lockfile was regenerated with the changes made
here.