build: update `poetry` to latest version
Poetry is the workflow management tool used for this project and forms
the root of all other actions taken when working with this repository.
It is also used to manage dependencies and therefore should be treated
very carefully, with updates to newer versions taken deliberately.

This PR updates `poetry` to the latest version of v2.1.4 to account for
the [changes introduced](https://python-poetry.org/history) in that bug
fix release, with the following actions taken:

* Bump all instances of `poetry` to the new version
  * Installs in workflows
  * Dockerfiles
  * pre-commit hook revision
  * Full comparison:
    * [python-poetry/poetry@2.1.3...2.1.4](https://github.com/python-poetry/poetry/compare/2.1.3...2.1.4)
* Bump dependencies in the lockfile to latest versions

None of the changes or new features in this new version required any
updates to the use of `poetry` in this project.