fix: use correct commit for tag pipelines
This change ensures the commit a tag points to is used for analysis
instead of the merge base between the tag commit and the default branch.
The `git merge-base` method only works when the tag is created with a
commit that lives on the default branch. Tags created elsewhere were
returning the incorrect SHA-1 sum.

Instead of attempting to find this commit ourselves, it turns out GitLab
does it for us and provides the value in the `CI_COMMIT_SHA` pre-defined
variable.

Additionally, log messages were cleaned up in a few spots to make it
more obvious when the `all_deps` property is set and a single commit is
used for analysis instead of a "common ancestor commit," which implies
a difference-based analysis.