feat: support tag pipelines in GitLab integration (#588)
This change updates the GitLab integration to allow for tag pipelines.
It also adds a prerequisite check to ensure that *only* explicitly
supported pipeline types are used to trigger the integration. This now
includes tag, branch, and merge request pipelines. Tag pipelines work
the same way as branch pipelines run against the default branch: *all*
dependencies are included in the analysis.

In the course of testing this change, a bug was discovered and fixed. It
was possible for an included manifest to force analysis *before* there
was a chance to set the `all_deps` property, causing tag or branch
pipelines running on the default branch to analyze only new deps instead
of all of them.

Specific changes made include:

* Ensure *all* dependencies are analyzed for tag pipelines and branch
pipelines run on the default branch
  * Reference `common_ancestor_commit` property in affected logic branch
* Create predicates to tell if operating within tag or branch pipelines
  * Use them to be more explicit about the logic paths taken
* Add prerequisite check to ensure operating within a supported pipeline
* Create new analysis label format for tag pipelines
  * Format: `GitLab-CI_tag_<tag_name>`
* Update documentation
  * Update links to GitLab documentation to the new, shorter form
* Format and refactor throughout
  * Present better debug log message when no new dependencies are found
  * Improve logs when single commit is used for analysis
  * Rename `is_in_mr` predicate to `is_in_mr_pipeline` for consistency
  * Extract to a function the process to acquire the git merge base

Closes [SFP-161](https://veracode.atlassian.net/browse/SFP-161)