ci: specify artifact download name and path
This change accounts for a breaking major version update in the GitHub
action `actions/download-artifact`, released three weeks ago.

The breaking change was inspected at the time and not deemed to affect
this repository. However, that does not appear to be the case since the
preview workflow is failing now in a way that makes it appear the
download was placed in the working directory instead of a new
subdirectory with the same name as the artifact. The working assumption
is that the action, when used with all default values, is behaving as if
artifact IDs were provided, which triggers the changed behavior in the
new version of the action.

This PR ensures that each use of the `download-artifact` action
specifies explicit `name` and `path` inputs for the artifact.

Ref: https://github.com/actions/download-artifact/releases/tag/v5.0.0