build: fix the release (#591)
This change reverts the previous release bump commit since the release
workflow failed. The reason it failed is due to the `pyproject.toml`
config file not getting updated correctly with the new version. This has
been corrected, by configuring Python Semantic Release (PSR) to update
the version in the new location.