build: update `poetry` to latest version (#598)
Poetry is the workflow management tool used for this project and forms
the root of all other actions taken when working with this repository.
It is also used to manage dependencies and therefore should be treated
very carefully, with updates to newer versions taken deliberately.

This PR updates `poetry` to the latest version of v2.2.1 to account for
the [changes introduced](https://python-poetry.org/history) in the
[minor version
release](https://python-poetry.org/blog/announcing-poetry-2.2.0/) and
the [bugfix version
release](https://github.com/python-poetry/poetry/releases/tag/2.2.1),
with these actions taken:

* Bump all instances of `poetry` to the new version
  * Installs in workflows
  * pre-commit hook revision
  * Dockerfiles
  * Full comparison:
    * https://github.com/python-poetry/poetry/compare/2.1.4...2.2.1
* Bump the `poetry-core` build system version in `pyproject.toml`
  * v2.2.1 matches Poetry version and adopts matching changes
  * Full comparison:
    * https://github.com/python-poetry/poetry-core/compare/2.1.2...2.2.1
* Update the lockfile with the new version of `poetry`

These updates changed the way `poetry` is used in this project. The
changes made include:

* The `dev` dependency group can be used to specify both `test` and `qa`
* Update `pyproject.toml` to match new expected PEP 735 compliant format

Other changes made include:

* Update documentation
  * `CONTRIBUTING.md` was brought current
* Update tests
  * Adjust for new expected `poetry-core` version
* Bump pre-commit hooks in `.pre-commit-config.yaml`